update: cleanup redundant code
