+++ 2. WhittedStyleBVH/renderer.h
@@ -1,8 +1,8 @@
 #pragma once
 
+#include "helper.h"
 #include "texture.h"
 #include "material.h"
-#include "model.h"
 #include "bvh_model.h"
 #include "grid_model.h"
 #include "hit_info.h"
