+++ 2. WhittedStyleBVH/scenes/base_scene.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include "model.h"
 #include "texture.h"
 #include "material.h"
 #include "bvh.h"
