+++ 2. WhittedStyleBVH/scenes/scene_1.h
@@ -21,8 +21,6 @@ namespace Tmpl8
 		int GetTriangleCount() const;
 	public:
 		float animTime = 0;
-		Model wok;
-		Model wok2;
 		//Model models[NUM_CUBE];
 		Texture skydome;
 		BVH sceneBVH;
