+++ 2. WhittedStyleBVH/whitted-style-bvh.vcxproj
@@ -136,7 +136,6 @@
     <ClCompile Include="..\lib\imgui\imgui_impl_opengl3.cpp" />
     <ClCompile Include="..\lib\imgui\imgui_tables.cpp" />
     <ClCompile Include="..\lib\imgui\imgui_widgets.cpp" />
-    <ClCompile Include="..\template\model.cpp" />
     <ClCompile Include="..\template\opencl.cpp" />
     <ClCompile Include="..\template\opengl.cpp" />
     <ClCompile Include="..\template\surface.cpp" />
@@ -150,7 +149,6 @@
     <ClCompile Include="grid.cpp" />
     <ClCompile Include="grid_model.cpp" />
     <ClCompile Include="renderer.cpp" />
-    <ClCompile Include="scenes\animable_scene.cpp" />
     <ClCompile Include="scenes\primitive_scene.cpp" />
     <ClCompile Include="scenes\scene_1.cpp" />
     <ClCompile Include="scenes\scene_2.cpp" />
@@ -168,7 +166,6 @@
     <ClInclude Include="..\template\camera.h" />
     <ClInclude Include="..\template\common.h" />
     <ClInclude Include="..\template\material.h" />
-    <ClInclude Include="..\template\model.h" />
     <ClInclude Include="..\template\opencl.h" />
     <ClInclude Include="..\template\opengl.h" />
     <ClInclude Include="..\template\precomp.h" />
@@ -185,7 +182,6 @@
     <ClInclude Include="helper.h" />
     <ClInclude Include="hit_info.h" />
     <ClInclude Include="renderer.h" />
-    <ClInclude Include="scenes\animable_scene.h" />
     <ClInclude Include="scenes\base_scene.h" />
     <ClInclude Include="scenes\primitive_scene.h" />
     <ClInclude Include="scenes\scene_1.h" />
